requirements.txt
loguru==0.6.0
Pillow==9.2.0
